requirements.txt
alembic==1.7.7
async-timeout==4.0.2
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
Flask==2.1.2
Flask-Assets==2.0
Flask-Compress==1.12
Flask-Migrate==3.1.0
flask-redis==0.4.0
Flask-SQLAlchemy==2.5.1
fmpsdk==20220405.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
plotly==5.8.0
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.36
tenacity==8.0.1
urllib3==1.26.9
webassets==2.0
Werkzeug==2.1.2
wrapt==1.14.1
